README.md
This is a README that describes the python code- it helps you understand the direction and steps taken to replicate the code for learning, advancement, adjustment, personalisation for future use and many others based known to you. 

This README contains a short description, getting started, running the code, breakdown of the code, deployment, author, license and acknowledgments.